requirements.txt
# --- bot ---
aiogram==3.25.0

# --- web ---
aiohttp==3.11.11

# --- config ---
environs==14.6.0

# --- logs ---
dlogger-drawiks==0.2.2
